+++ apollo-normalized-cache-common/src/commonMain/kotlin/com/apollographql/apollo3/cache/normalized/internal/OperationCacheExtensions.kt
@@ -1,77 +1,50 @@
 package com.apollographql.apollo3.cache.normalized.internal
 
-import com.apollographql.apollo3.api.Executable
 import com.apollographql.apollo3.api.CustomScalarAdapters
+import com.apollographql.apollo3.api.Executable
 import com.apollographql.apollo3.api.Fragment
 import com.apollographql.apollo3.api.Operation
-import com.apollographql.apollo3.cache.normalized.Record
 import com.apollographql.apollo3.api.internal.json.MapJsonReader
 import com.apollographql.apollo3.api.internal.json.MapJsonWriter
-import com.apollographql.apollo3.api.Adapter
-import com.apollographql.apollo3.api.CompiledSelection
 import com.apollographql.apollo3.api.variables
 import com.apollographql.apollo3.cache.CacheHeaders
 import com.apollographql.apollo3.cache.normalized.CacheKey
 import com.apollographql.apollo3.cache.normalized.CacheResolver
 import com.apollographql.apollo3.cache.normalized.ReadOnlyNormalizedCache
+import com.apollographql.apollo3.cache.normalized.Record
+
 
 fun <D : Operation.Data> Operation<D>.normalize(
     data: D,
     customScalarAdapters: CustomScalarAdapters,
     cacheResolver: CacheResolver,
-) = normalizeInternal(
-    data,
-    customScalarAdapters,
-    cacheResolver,
-    CacheKey.rootKey().key,
-    adapter(),
-    variables(customScalarAdapters),
-    selections())
+) = normalize(data, customScalarAdapters, cacheResolver, CacheKey.rootKey().key)
 
-fun <D : Fragment.Data> Fragment<D>.normalize(
+@Suppress("UNCHECKED_CAST")
+fun <D : Executable.Data> Executable<D>.normalize(
     data: D,
     customScalarAdapters: CustomScalarAdapters,
     cacheResolver: CacheResolver,
     rootKey: String,
-) = normalizeInternal(
-    data,
-    customScalarAdapters,
-    cacheResolver,
-    rootKey,
-    adapter(),
-    variables(customScalarAdapters),
-    selections())
-
-private fun <D> normalizeInternal(
-    data: D,
-    customScalarAdapters: CustomScalarAdapters,
-    cacheResolver: CacheResolver,
-    rootKey: String,
-    adapter: Adapter<D>,
-    variables: Executable.Variables,
-    selections: List<CompiledSelection>,
 ): Map<String, Record> {
   val writer = MapJsonWriter()
-  adapter.toJson(writer, customScalarAdapters, data)
-  return Normalizer(variables, rootKey) { type, fields ->
+  adapter().toJson(writer, customScalarAdapters, data)
+  return DefaultNormalizer(variables(customScalarAdapters), rootKey) { type, fields ->
     cacheResolver.cacheKeyForObject(type, fields)?.key
-  }.normalize(writer.root() as Map<String, Any?>, selections)
+  }.normalize(writer.root() as Map<String, Any?>, selections())
 }
 
-fun <D : Operation.Data> Operation<D>.readDataFromCache(
+fun <D : Executable.Data> Executable<D>.readDataFromCache(
     customScalarAdapters: CustomScalarAdapters,
     cache: ReadOnlyNormalizedCache,
     cacheResolver: CacheResolver,
     cacheHeaders: CacheHeaders,
 ) = readInternal(
+    cacheKey = CacheKey.rootKey(),
+    customScalarAdapters = customScalarAdapters,
     cache = cache,
     cacheResolver = cacheResolver,
     cacheHeaders = cacheHeaders,
-    variables = variables(customScalarAdapters),
-    adapter = adapter(),
-    customScalarAdapters = customScalarAdapters,
-    cacheKey = CacheKey.rootKey(),
-    selections = selections()
 )
 
 fun <D : Fragment.Data> Fragment<D>.readDataFromCache(
@@ -82,39 +55,32 @@ fun <D : Fragment.Data> Fragment<D>.readDataFromCache(
     cacheHeaders: CacheHeaders,
 ) = readInternal(
     cacheKey = cacheKey,
+    customScalarAdapters = customScalarAdapters,
     cache = cache,
     cacheResolver = cacheResolver,
     cacheHeaders = cacheHeaders,
-    variables = variables(customScalarAdapters),
-    adapter = adapter(),
-    customScalarAdapters = customScalarAdapters,
-    selections = selections()
 )
 
-
-private fun <D> readInternal(
+private fun <D : Executable.Data> Executable<D>.readInternal(
     cacheKey: CacheKey,
+    customScalarAdapters: CustomScalarAdapters,
     cache: ReadOnlyNormalizedCache,
     cacheResolver: CacheResolver,
     cacheHeaders: CacheHeaders,
-    variables: Executable.Variables,
-    adapter: Adapter<D>,
-    customScalarAdapters: CustomScalarAdapters,
-    selections: List<CompiledSelection>,
-): D? {
+): D {
   val map = CacheBatchReader(
       cache = cache,
       cacheHeaders = cacheHeaders,
       cacheResolver = cacheResolver,
-      variables = variables,
+      variables = variables(customScalarAdapters),
       rootKey = cacheKey.key,
-      rootSelections = selections
+      rootSelections = selections()
   ).toMap()
 
   val reader = MapJsonReader(
       root = map,
   )
-  return adapter.fromJson(reader, customScalarAdapters)
+  return adapter().fromJson(reader, customScalarAdapters)
 }
 
 fun Collection<Record>?.dependentKeys(): Set<String> {
